Move sync-file-settings to notification subfolder
