+++ editors/code/src/lsp.ts
@@ -1,9 +1,10 @@
 import * as vscode from "vscode";
 import * as lc from "vscode-languageclient/node";
 import { default as PQueue } from "p-queue";
-import { getInitializationOptions, TomlSettings } from "./settings";
+import { getInitializationOptions } from "./settings";
+import { FileSettingsState } from "./notification/sync-file-settings";
 import { Middleware, ResponseError } from "vscode-languageclient/node";
-import { SYNC_FILE_SETTINGS } from "./lsp-ext";
+import { SYNC_FILE_SETTINGS } from "./notification/sync-file-settings";
 
 // All session management operations are put on a queue. They can't run
 // concurrently and either result in a started or stopped state. Starting when
@@ -24,13 +25,13 @@ export class Lsp {
 	private state = State.Stopped;
 	private stateQueue: PQueue;
 
-	private tomlSettings: TomlSettings;
+	private tomlSettings: FileSettingsState;
 
 	constructor(context: vscode.ExtensionContext) {
 		this.channel = vscode.window.createOutputChannel("Air Language Server");
 		context.subscriptions.push(this.channel);
 		this.stateQueue = new PQueue({ concurrency: 1 });
-		this.tomlSettings = new TomlSettings(context);
+		this.tomlSettings = new FileSettingsState(context);
 	}
 
 	public getClient(): lc.LanguageClient {
